Update README.md
Movie List App

This is a simple React app that allows you to manage a list of movies. The app uses hooks and props to manage the state of the movie list and allows you to add movies to the list, as well as sort them by name or rating.

Features
Add a new movie to the list
Sort the movies by name or rating
View the list of movies with their details, such as title, rating, and release date

Installation
To install and run the app, follow these steps:

Clone the repository to your local machine.
Open a terminal window and navigate to the root directory of the project.
Run the following command to install the necessary dependencies:

npm install

Once the dependencies are installed, start the app by running the following command:

npm start

The app should now be running at http://localhost:3000.